idpf: Refactor vport config operations in new file
Added new file 'idpf_vport_config.c'.

It moves vport config functions in new file 'idpf_vport_config.c'.
The purpose is to separate the vport config functionality into
separate file.

There is no change in the code or functionality before and after
the change.
